requirements.txt
slackclient>=1.0.5
slackeventsapi>=0.1
pyyaml>=3.12
SQLAlchemy>=0.7.9
